# files/E-Commerce OpenAPI Specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E-Commerce API",
    "description": "\nWelcome to the Mock E-Commerce API! 🚀\n\nThis API provides a comprehensive set of functionalities for managing your e-commerce platform.\nKey features include:\n\n- **CRUD**\n\t- Create, Read, Update, and Delete endpoints.\n- **Search**\n\t- Find specific information with parameters and pagination.\n- **Auth**\n\t- Verify user/admin identity.\n\t- Secure with Access and Refresh tokens.\n- **Permission**\n\t- Assign roles with specific permissions.\n\t- Different access levels for User/Admin.\n- **Validation**\n\t- Ensure accurate and secure input data.\n\nFor any inquiries, please contact sshah@watermelon.us\n",
    "contact": {
      "name": "Sahil Shah",
      "url": "https://github.com/WM-SahilShah"
    },
    "version": "2.0.4"
  },
  "paths": {
...
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get All Products",
        "description": "This endpoint retrieves all products with pagination (required) and search by title (optional).",
        "operationId": "get_all_products_products__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (Required)",
              "default": "<integer>*",
              "title": "Page"
            },
            "description": "Page number (Required)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page (Required)",
              "default": "<integer>*",
              "title": "Limit"
            },
            "description": "Items per page (Required)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Search based title of products",
              "default": "<string>",
              "title": "Search"
            },
            "description": "Search based title of products"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create New Product ##",
        "description": "This endpoint allows an admin to create a new product.",
        "operationId": "create_product_products__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get Specific Product",
        "description": "This endpoint retrieves a specific product by its ID.",
        "operationId": "get_product_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update Existing Product ##",
        "description": "This endpoint allows an admin to update an existing product by its ID.",
        "operationId": "update_product_products__product_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete Existing Category ##",
        "description": "Delete a product by its ID.",
        "operationId": "delete_product_products__product_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOutDelete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProductBase": {
        "properties": {
          "id": {
            "type": "integer",
            "maximum": 299,
            "minimum": 200,
            "title": "Id",
            "description": "Unique identifier for the product.",
            "default": "<integer>"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "Title of the product.",
            "default": "<string>"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description",
            "description": "Description of the product.",
            "default": "<string>"
          },
          "price": {
            "type": "integer",
            "minimum": 0,
            "title": "Price",
            "description": "Price of the product (>=0).",
            "default": "<integer>"
          },
          "discount_percentage": {
            "type": "number",
            "maximum": 100,
            "minimum": 0,
            "title": "Discount Percentage",
            "description": "Discount percentage (0-100).",
            "default": "<float>"
          },
          "rating": {
            "type": "number",
            "maximum": 5,
            "minimum": 0,
            "title": "Rating",
            "description": "Rating of the product (0-5).",
            "default": "<float>"
          },
          "stock": {
            "type": "integer",
            "minimum": 0,
            "title": "Stock",
            "description": "Stock count (>=0).",
            "default": "<integer>"
          },
          "brand": {
            "type": "string",
            "minLength": 1,
            "title": "Brand",
            "description": "Brand of the product.",
            "default": "<string>"
          },
          "thumbnail": {
            "type": "string",
            "minLength": 1,
            "title": "Thumbnail",
            "description": "URL of the product thumbnail.",
            "default": "<url>"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Images",
            "description": "List of image URLs for the product.",
            "default": "<url>"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published",
            "description": "Whether the product is published or not.",
            "default": "<boolean>"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the product was created.",
            "default": "<datetime obj/ISO 8601 string>"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id",
            "description": "Unique identifier for the product's category.",
            "default": "<integer>"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "price",
          "discount_percentage",
          "rating",
          "stock",
          "brand",
          "thumbnail",
          "images",
          "is_published",
          "created_at",
          "category_id"
        ],
        "title": "ProductBase",
        "description": "Represents basic details of a product.\n\nAttributes:\n- `id` (int): Unique identifier for the product.\n- `title` (str): Title of the product.\n- `description` (str): Description of the product.\n- `price` (int): Price of the product (>= 0).\n- `discount_percentage` (float): Discount percentage (0-100).\n- `rating` (float): Rating of the product (0-5).\n- `stock` (int): Stock count (>= 0).\n- `brand` (str): Brand of the product.\n- `thumbnail` (str): URL of the product thumbnail.\n- `images` (List[str]): List of image URLs for the product.\n- `is_published` (bool): Whether the product is published or not.\n- `created_at` (datetime): Timestamp when the product was created.\n- `category_id` (int): Unique identifier for the product's category."
      },
      "ProductCreate": {
        "properties": {
          "id": {
            "type": "integer",
            "maximum": 299,
            "minimum": 200,
            "title": "Id",
            "description": "Unique identifier for the product.",
            "default": "<integer>"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "Title of the product.",
            "default": "<string>"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description",
            "description": "Description of the product.",
            "default": "<string>"
          },
          "price": {
            "type": "integer",
            "minimum": 0,
            "title": "Price",
            "description": "Price of the product (>=0).",
            "default": "<integer>"
          },
          "discount_percentage": {
            "type": "number",
            "maximum": 100,
            "minimum": 0,
            "title": "Discount Percentage",
            "description": "Discount percentage (0-100).",
            "default": "<float>"
          },
          "rating": {
            "type": "number",
            "maximum": 5,
            "minimum": 0,
            "title": "Rating",
            "description": "Rating of the product (0-5).",
            "default": "<float>"
          },
          "stock": {
            "type": "integer",
            "minimum": 0,
            "title": "Stock",
            "description": "Stock count (>=0).",
            "default": "<integer>"
          },
          "brand": {
            "type": "string",
            "minLength": 1,
            "title": "Brand",
            "description": "Brand of the product.",
            "default": "<string>"
          },
          "thumbnail": {
            "type": "string",
            "minLength": 1,
            "title": "Thumbnail",
            "description": "URL of the product thumbnail.",
            "default": "<url>"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Images",
            "description": "List of image URLs for the product.",
            "default": "<url>"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published",
            "description": "Whether the product is published or not.",
            "default": "<boolean>"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the product was created.",
            "default": "<datetime obj/ISO 8601 string>"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id",
            "description": "Unique identifier for the product's category.",
            "default": "<integer>"
          }
        },
        "type": "object",
        "title": "ProductCreate",
        "description": "Represents the schema for creating a product.\n\nAttributes:\n- `id` (int): Unique identifier for the product.\n- `title` (str): Title of the product.\n- `description` (str): Description of the product.\n- `price` (int): Price of the product (>= 0).\n- `discount_percentage` (float): Discount percentage (0-100).\n- `rating` (float): Rating of the product (0-5).\n- `stock` (int): Stock count (>= 0).\n- `brand` (str): Brand of the product.\n- `thumbnail` (str): URL of the product thumbnail.\n- `images` (List[str]): List of image URLs for the product.\n- `is_published` (bool): Whether the product is published or not.\n- `created_at` (datetime): Timestamp when the product was created.\n- `category_id` (int): Unique identifier for the product's category."
      },
      "ProductOut": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message."
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductBase"
              }
            ],
            "description": "Product details."
          }
        },
        "type": "object",
        "required": [
          "message",
          "data"
        ],
        "title": "ProductOut",
        "description": "Represents the output schema for a single product.\n\nAttributes:\n- `message` (str): Response message.\n- `data` (ProductBase): Details of the product.\n    - `id` (int): Unique identifier for the product.\n    - `title` (str): Title of the product.\n    - `description` (str): Description of the product.\n    - `price` (int): Price of the product (>= 0).\n    - `discount_percentage` (float): Discount percentage (0-100).\n    - `rating` (float): Rating of the product (0-5).\n    - `stock` (int): Stock count (>= 0).\n    - `brand` (str): Brand of the product.\n    - `thumbnail` (str): URL of the product thumbnail.\n    - `images` (List[str]): List of image URLs for the product.\n    - `is_published` (bool): Whether the product is published or not.\n    - `created_at` (datetime): Timestamp when the product was created.\n    - `category_id` (int): Unique identifier for the product's category."
      },
      "ProductOutDelete": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message."
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductBase"
              }
            ],
            "description": "Product details."
          }
        },
        "type": "object",
        "required": [
          "message",
          "data"
        ],
        "title": "ProductOutDelete",
        "description": "Represents the output schema for product deletion.\n\nAttributes:\n- `message` (str): Response message.\n- `data` (ProductBase): Details of the deleted product.\n    - `id` (int): Unique identifier for the product.\n    - `title` (str): Title of the product.\n    - `description` (str): Description of the product.\n    - `price` (int): Price of the product (>= 0).\n    - `discount_percentage` (float): Discount percentage (0-100).\n    - `rating` (float): Rating of the product (0-5).\n    - `stock` (int): Stock count (>= 0).\n    - `brand` (str): Brand of the product.\n    - `thumbnail` (str): URL of the product thumbnail.\n    - `images` (List[str]): List of image URLs for the product.\n    - `is_published` (bool): Whether the product is published or not.\n    - `created_at` (datetime): Timestamp when the product was created.\n    - `category_id` (int): Unique identifier for the product's category."
      },
      "ProductUpdate": {
        "properties": {
          "id": {
            "type": "integer",
            "maximum": 299,
            "minimum": 200,
            "title": "Id",
            "description": "Unique identifier for the product.",
            "default": "<integer>"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "Title of the product.",
            "default": "<string>"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description",
            "description": "Description of the product.",
            "default": "<string>"
          },
          "price": {
            "type": "integer",
            "minimum": 0,
            "title": "Price",
            "description": "Price of the product (>=0).",
            "default": "<integer>"
          },
          "discount_percentage": {
            "type": "number",
            "maximum": 100,
            "minimum": 0,
            "title": "Discount Percentage",
            "description": "Discount percentage (0-100).",
            "default": "<float>"
          },
          "rating": {
            "type": "number",
            "maximum": 5,
            "minimum": 0,
            "title": "Rating",
            "description": "Rating of the product (0-5).",
            "default": "<float>"
          },
          "stock": {
            "type": "integer",
            "minimum": 0,
            "title": "Stock",
            "description": "Stock count (>=0).",
            "default": "<integer>"
          },
          "brand": {
            "type": "string",
            "minLength": 1,
            "title": "Brand",
            "description": "Brand of the product.",
            "default": "<string>"
          },
          "thumbnail": {
            "type": "string",
            "minLength": 1,
            "title": "Thumbnail",
            "description": "URL of the product thumbnail.",
            "default": "<url>"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Images",
            "description": "List of image URLs for the product.",
            "default": "<url>"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published",
            "description": "Whether the product is published or not.",
            "default": "<boolean>"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the product was created.",
            "default": "<datetime obj/ISO 8601 string>"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id",
            "description": "Unique identifier for the product's category.",
            "default": "<integer>"
          }
        },
        "type": "object",
        "title": "ProductUpdate",
        "description": "Represents the schema for updating a product.\n\nAttributes:\n- `id` (int): Unique identifier for the product.\n- `title` (str): Title of the product.\n- `description` (str): Description of the product.\n- `price` (int): Price of the product (>= 0).\n- `discount_percentage` (float): Discount percentage (0-100).\n- `rating` (float): Rating of the product (0-5).\n- `stock` (int): Stock count (>= 0).\n- `brand` (str): Brand of the product.\n- `thumbnail` (str): URL of the product thumbnail.\n- `images` (List[str]): List of image URLs for the product.\n- `is_published` (bool): Whether the product is published or not.\n- `created_at` (datetime): Timestamp when the product was created.\n- `category_id` (int): Unique identifier for the product's category."
      },
      "ProductsOut": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message."
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProductBase"
            },
            "type": "array",
            "title": "Data",
            "description": "List of product details."
          }
        },
        "type": "object",
        "required": [
          "message",
          "data"
        ],
        "title": "ProductsOut",
        "description": "Represents the output schema for multiple products.\n\nAttributes:\n- `message` (str): Response message.\n- `data` (List[ProductBase]): List of product details.\n    - `id` (int): Unique identifier for the product.\n    - `title` (str): Title of the product.\n    - `description` (str): Description of the product.\n    - `price` (int): Price of the product (>= 0).\n    - `discount_percentage` (float): Discount percentage (0-100).\n    - `rating` (float): Rating of the product (0-5).\n    - `stock` (int): Stock count (>= 0).\n    - `brand` (str): Brand of the product.\n    - `thumbnail` (str): URL of the product thumbnail.\n    - `images` (List[str]): List of image URLs for the product.\n    - `is_published` (bool): Whether the product is published or not.\n    - `created_at` (datetime): Timestamp when the product was created.\n    - `category_id` (int): Unique identifier for the product's category."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}